Clase HTTP y simplificacion de solicitudes

+++ admin/crud_users/crud.php
@@ -1,12 +1,13 @@
 <?php
+session_start();
+include '../../config.php';
+include '../../models/Http.php';
 if (isset($_GET['success']) && $_GET['success']) {
     echo '<script>alert("Desactivado correctamente")</script>';
 }
-session_start();
-if (isset($_SESSION['id_admin'], $_SESSION['username'], $_SESSION['email'], $_SESSION['token'])) {
+if (isset($_SESSION['id_admin'], $_SESSION['username'], $_SESSION['token'])) {
     $id_admin = $_SESSION['id_admin'];
     $username = $_SESSION['username'];
-    $email = $_SESSION['email'];
     $token = $_SESSION['token'];
 } else {
     header("Location: ../../catalogo/login.php");
@@ -22,6 +23,8 @@ if (isset($_SESSION['id_admin'], $_SESSION['username'], $_SESSION['email'], $_SE
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma-rtl.min.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.2/css/all.min.css">
+    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+
     <link rel="stylesheet" href="./estilos/crud.css">
     <title>Usuarios</title>
 </head>
@@ -77,61 +80,48 @@ if (isset($_SESSION['id_admin'], $_SESSION['username'], $_SESSION['email'], $_SE
             </thead>
             <tbody>
                 <?php
-                require '../../database/conexion.php';
+                
 
-                $database = new Database;
-                $conexion = $database->connect();
 
-                $sql = "SELECT c.id, tp.id AS tipo_id, c.identificacion, c.primerNombre, c.segundoNombre, c.primerApellido, c.segundoApellido, c.telefono, c.email 
-        FROM usuarios AS c 
-        LEFT JOIN tiposid AS tp ON tp.codId = c.tipoId
-        WHERE c.activo = 1";
+                $url = URL."/usuarios";
 
-                $parameters = [];
-
+                $params = [];
                 if (isset($_GET['name']) && !empty($_GET['name'])) {
-                    $name = '%' . $_GET['name'] . '%';
-
-                    $sql .= " AND (
-        c.primerNombre LIKE :name OR 
-        c.segundoNombre LIKE :name OR 
-        c.primerApellido LIKE :name OR
-        c.segundoApellido LIKE :name OR
-        c.email LIKE :name OR
-        c.telefono LIKE :name
-    )";
-
-                    $parameters[':name'] = $name;
+                    $params['pv'] = urlencode($_GET['name']);
+                }
+                if (isset($_GET['type']) && $_GET['type'] != 0) {
+                    $params['tp'] = $_GET['type'];
                 }
 
-                if (isset($_GET['type']) && is_numeric($_GET['type'])) {
-                    $tipoId = $_GET['type'];
+                if (!empty($params)) {
+                    $url .= '?' . http_build_query($params);
+                }
 
-                    if ($tipoId != 0) {
-                        $sql .= " AND c.tipoId = :tipoId";
-                        $parameters[':tipoId'] = $tipoId;
+                HttpClient::setUrl($url);
+                $usuarios = HttpClient::get();
+                
+                if (empty($usuarios['usuarios'])) {
+                    $nullable = true;
+                } else {
+                    foreach ($usuarios['usuarios'] as $key => $row) {
+                        echo '<tr>';
+                        echo '<td>' . $row["id"] . '</td>';
+                        echo '<td>' . $row["tipoid"]['id'] . '</td>';
+                        echo '<td>' . $row["identificacion"] . '</td>';
+                        echo '<td>' . $row["primerNombre"] . '</td>';
+                        echo '<td>' . $row["segundoNombre"] . '</td>';
+                        echo '<td>' . $row["primerApellido"] . '</td>';
+                        echo '<td>' . $row["segundoApellido"] . '</td>';
+                        echo '<td>' . $row["email"] . '</td>';
+                        echo '<td>' . $row["telefono"] . '</td>';
+                        echo '<td><a href="update.php?id=' . $row["id"] . '" class="button is-link">Editar</a> <a href="delete.php?id=' . $row["id"] . '" class="button is-danger">Eliminar</a></td>';
+                        echo '</tr>';
+                        if(isset($nullable) && $nullable){
+                            echo '<td colspan=10>No encontrado</td>';
+                        }
                     }
                 }
 
-                $result = $conexion->prepare($sql);
-                $result->execute($parameters);
-
-                $usuarios = $result->fetchAll(PDO::FETCH_ASSOC);
-
-                foreach ($usuarios as $row) {
-                    echo '<tr>';
-                    echo '<td>' . $row["id"] . '</td>';
-                    echo '<td>' . $row["tipo_id"] . '</td>';
-                    echo '<td>' . $row["identificacion"] . '</td>';
-                    echo '<td>' . $row["primerNombre"] . '</td>';
-                    echo '<td>' . $row["segundoNombre"] . '</td>';
-                    echo '<td>' . $row["primerApellido"] . '</td>';
-                    echo '<td>' . $row["segundoApellido"] . '</td>';
-                    echo '<td>' . $row["email"] . '</td>';
-                    echo '<td>' . $row["telefono"] . '</td>';
-                    echo '<td><a href="update.php?id=' . $row["id"] . '" class="button is-link">Editar</a> <a href="delete.php?id=' . $row["id"] . '" class="button is-danger">Eliminar</a></td>';
-                    echo '</tr>';
-                }
                 ?>
 
             </tbody>
